13/3/2022 (#70)
* Favor system and Command Prompt

* Update CommandScene.unity

* Input

* Fixing the small bugs of our game

* Favor system and Command Prompt

* Update CommandScene.unity

* Input of each command set working

* Update Electronic Highway Sign SDF.asset

* fixing input field while opening and close console

* bug fix: activate save process UI animation

* Dev Menu

* Update EditorBuildSettings.asset

* Update EditorUserSettings.asset

* input command handles invalid  inputs

Handles invalid inputs and displays the correct handler onto screen as well as follows the procedure of shutting down if there is too many attempts

* pausing and deactivate input during processing

* moving it to the game scene

* Terminal Ui scaling

* Testing

* BUGSSS

* One shot animation and two animations state scripts

* All command set up and processing help

* saving terminal

* Update CommandScene.unity

* Tutorial Dialogues

* testing

* allow user to change name and make invalid inputs for all commands

* updating terminal prefab

* Favor

* fix

* Favour System | Enemy Bug

* Tutorial

* Testing

* allowing a press any key functionality (not letting actual typing during this section)

* quick changes on displayment text

* New texture replace

* Upload more texture

* Hiding transition | Minor Bug

* Update GameCommandPrompt.cs

* importing text font and testing it

* Update GameTesting.unity

* STOP | REWIND

* UI Scaling Fix

* Main Menu muzic

* Main Menu Muzik

* fixing The font decision and fixing the door problem

* New materials

* import UI and updating UI on menu

* Update MainMenu.unity

* Bench and Sink

* Fixing volume problems and scene

* Setting up a third ability and randomising

* Update EditorUserSettings.asset

* Clock hand rotation

* Rewind

* Dialagues

* Piano and Stage Update

* new folder

* here

* test

* textbox UI alignment, font size

* Wheelchair Modern update

* mask

* testing audio name tag

* Fixxes

* Modern Hospital

* 1870s materials Update

* fix small bugs

* player position fixing,narrator voice and menu

* fixing respawn

* Enemy Kill

* More materials

* toggle on favor points and timer for objective

* Coffee Pot

* testing more for opening and closing drawers (moving position of child object)

* import future door, TV, couch

* Update Electronic Highway Sign SDF.asset

* Kitchen

* here

* Update Electronic Highway Sign SDF.asset

* ugh

* PC crashing 😑

* Hiding Location | Enemy Safe Rooom

* Maskk

* Mask Init

* Mask Change Done

* removing amulet and teleport on favor system

* ObjTut Bug

* Start Game Bugs

* yee

* edge case

* Hospital and Kitchen

* Notication UI finally done

* kill animations imported

Rine/enemy/Enemy KillAnims

* hiding location fix

* Futuristic Kitchen Counter

* Fixes

* hiding spots prefab

* Update Path_face 3.unity

* door VFX start and changed hiding spot indicator

* PC Crashed before i could save changes!!!

* Replacing the assets

+++ Assets/Scripts/rachael/SceneChanger.cs
@@ -12,13 +12,10 @@ namespace rachael
         // Start is called before the first frame update
         void Start()
         {
+        
             if(m_continueBlock != null)
             {
-                if (!PlayerPrefs.HasKey("SaveFile"))
-                {
-                    m_continueBlock.SetActive(false);
-
-                }
+                ButtonPresent();
 
             }
         }
@@ -33,15 +30,17 @@ namespace rachael
         {
             if (Directory.Exists(Application.persistentDataPath + "/player/"))
             {
-                m_continueBlock.SetActive(false);
+                m_continueBlock.SetActive(true);
             }
             else
             {
-                m_continueBlock.SetActive(true);
+                m_continueBlock.SetActive(false);
             }
         }
         public void MainScene()
         {
+            Cursor.visible = true;
+
             Cursor.lockState = CursorLockMode.None;
             SceneManager.LoadScene(0);
         }
@@ -67,13 +66,18 @@ namespace rachael
         public void Ending()
         {
             Cursor.lockState = CursorLockMode.None;
+            Cursor.visible = true;
             SceneManager.LoadScene(2);
         }
         public void DeleteFile()
         {
-            PlayerPrefs.DeleteAll();
-            SaveSystem.SaveSystem.DeleteNarrator();
-            SaveSystem.SaveSystem.DeletePlayer();
+
+            if (Directory.Exists(Application.persistentDataPath + "/player/"))
+                SaveSystem.SaveSystem.DeletePlayer();
+            if (Directory.Exists(Application.persistentDataPath + "/narrator/"))
+                SaveSystem.SaveSystem.DeleteNarrator();
+
+            ButtonPresent();
             
         }
         //public void Continue()
@@ -90,6 +94,7 @@ namespace rachael
 
         public void LoseScreen()
         {
+            Cursor.visible = true;
             Cursor.lockState = CursorLockMode.None;
             SceneManager.LoadScene(3);
